pay_term_config/requirements.txt
Flask==3.0.0
mfrc522-python==0.0.5
python-dotenv==1.0.0
spidev==3.6
RPi.GPIO==0.7.1
web3==6.10.0